Refactor json converter
Move json classes into separate files.
