+++ lib/converters/json/document-creation-info.ts
@@ -1,0 +1,28 @@
+import type { Document } from "../../spdx2model/document";
+import { formatDatetime } from "./utils";
+
+export class JsonDocumentCreationInfo {
+  created: string;
+  creators: string[];
+  comment?: string;
+  licenseListVersion?: string;
+
+  constructor(
+    created: string,
+    creators: string[],
+    comment?: string,
+    licenseListVersion?: string,
+  ) {
+    this.created = created;
+    this.creators = creators;
+    this.comment = comment;
+    this.licenseListVersion = licenseListVersion;
+  }
+
+  static fromDocument(document: Document): JsonDocumentCreationInfo {
+    return new JsonDocumentCreationInfo(
+      formatDatetime(document.creationInfo.created),
+      document.creationInfo.creators,
+    );
+  }
+}
